note about the script
This script is based on a module called itchat. It could sign in by your Wechat account and it can give you some interesting information about your friends, such as the provincial distribution, the "word cloud" of your friends' signature.


When you run this script, some errors may occur. That may be due to the fonts. Because most information in Wechat is written in Chinese, while my computer system does not have the fonts which Matplotlib needs, I downloaded them manually. When you run this script and have some errors, check that first.